api_ai/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.0
blessed==1.20.0
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
distro==1.9.0
Django==5.0.8
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-compat==1.0.15
django-picklefield==3.2
django-timezone-field==7.0
djangorestframework==3.15.2
dnspython==2.6.1
eventlet==0.36.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
jinxed==1.3.0
jiter==0.5.0
kombu==5.4.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-stdnum==1.20
redis==3.5.3
requests==2.32.3
rsa==4.9
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
